Refactoring (#43)
